pyproject.toml
[tool.poetry]
name = "xrays"
version = "0.1.0"
description = ""
authors = ["Johannes Wienke <languitar@semipol.de>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
pandas = "^1.0.4"
plotly = "^4.8.1"
tqdm = "^4.46.1"
click-pathlib = "^2020.3.13"
fastparquet = { version = "^0.4.0", extras = ["snappy"] }
packaging = "^20.4"
natsort = "^7.0.1"
dash = "^1.12.0"
gitpython = "^3.1.8"

[tool.poetry.dev-dependencies]
mypy = "^0.780"
flake8 = "^3.8.3"
black = "^19.10b0"
ipython = "^7.15.0"
pytest = "^6.0.1"
pytest-sugar = "^0.9.4"
pytest-git = "^1.7.0"

[tool.poetry.scripts]
xrays = 'xrays:hotspots'

[tool.pytest.ini_options]
log_level = 'DEBUG'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
